backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# OpenAI APIs
openai==1.54.4
openai-agents==0.0.14

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# File handling
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
markdown==3.5.1

# Real-time communication
python-socketio==5.10.0

# HTTP requests
httpx==0.25.2
aiofiles==23.2.1

# Data validation
pydantic==2.5.0
pydantic-settings==2.0.3

# Background tasks
celery==5.3.4
redis==5.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
